Moved editor files
